chore(v3): clean up spfx v3 plugin code (#5916)
* feat: up

* fix: up
